Always use https in production.
This is too important to risk screwing up. Misconfigured production
settings could cause the scheme to be http and the credentials to be
leaked.

In a well-configured production environment, this shouldn't make a
difference because the request.scheme should be https. If a front-end
server is proxying to django in plaintext, it should set the
X-Forwarded-Proto header and the django setting `SECURE_PROXY_SSL_HEADER
= ('HTTP_X_FORWARDED_PROTO', 'https')` should be set so that it is
trusted.

I'm slightly reluctant to make this change though because there is the
valid use case of using DEBUG=False without using tls in a secure
development or staging environment. However, in my opinion it's easier
and safer to use self-signed certs in these environments than to hack
the https out of everything. And regardless, the only ill effect of this
change in this circumstance is that the emailed url will be broken until
you delete the 's'.